Refresh the Virginia COVID-19 Data.

Updates the Virginia COVID-19 cases.

refresh_VirginiaC19(verbose = TRUE)

Arguments

verbose
Should messages be displayed?

Source

VirginiaC19

covid19R documentation

Value

A tibble object

date - The date in YYYY-MM-DD form

location - The name of the location as provided by the data source.

location_type - The type of location using the covid19R controlled vocabulary.

location_code - A standardized location code using a national or international standard.

location_code_type The type of standardized location code being used according to the covid19R controlled vocabulary.

data_type - the type of data in that given row using the covid19R controlled vocabulary. Daily totals are found here.

value - number of cases of each data type

Details

This package scrapes the state of Virginia reported cases.

Examples

if (FALSE) {
refresh_VirginiaC19()
}